repos/SBS-Backend-main/python/playoffs-metadata-api/requirements.txt
# Function dependencies, for example:
# package>=version
# Function dependencies, for example:
# package>=version
google-cloud-firestore==2.0.2